System.Metrics.Label

Description

This module defines a type for mutable, string-valued labels. Labels are variable values and can be used to track e.g. the command line arguments or other free-form values. All operations on labels are thread-safe.

Documentation

data Label

A mutable, text-valued label.

new :: IO Label

Create a new empty label.

read :: Label -> IO Text

Get the current value of the label.

set :: Label -> Text -> IO ()

Set the label to the given value.

modify :: (Text -> Text) -> Label -> IO ()

Set the label to the result of applying the given function to the value.